Move whole code to listeners module
