+++ kinto_signer/__init__.py
@@ -1,202 +1,16 @@
 import pkg_resources
 import functools
 
-import transaction
-from kinto.core import errors
 from kinto.core.events import ACTIONS, ResourceChanged
-from kinto import logger
-from pyramid import httpexceptions
 from pyramid.exceptions import ConfigurationError
 from pyramid.settings import asbool
 
-from kinto_signer.updater import LocalUpdater
 from kinto_signer.signer import heartbeat
 from kinto_signer import utils
+from kinto_signer import listeners
 
 #: Module version, as defined in PEP-0396.
 __version__ = pkg_resources.get_distribution(__package__).version
-
-
-def raise_invalid(**kwargs):
-    transaction.doom()
-    raise errors.http_error(httpexceptions.HTTPBadRequest(), **kwargs)
-
-
-def raise_forbidden(**kwargs):
-    transaction.doom()
-    raise errors.http_error(httpexceptions.HTTPForbidden(), **kwargs)
-
-
-def sign_collection_data(event, resources):
-    """
-    Listen to resource change events, to check if a new signature is
-    requested.
-
-    When a source collection specified in settings is modified, and its
-    new metadata ``status`` is set to ``"to-sign"``, then sign the data
-    and update the destination.
-    """
-    payload = event.payload
-
-    if 'bucket_id' not in payload:
-        # Safety check for kinto < 3.3 where events have incoherent payloads
-        # on default bucket.
-        return
-
-    for impacted in event.impacted_records:
-        new_collection = impacted['new']
-
-        key = "/buckets/{bucket_id}/collections/{collection_id}".format(
-            collection_id=new_collection['id'],
-            bucket_id=payload['bucket_id'])
-
-        resource = resources.get(key)
-
-        # Only sign the configured resources.
-        if resource is None:
-            continue
-
-        registry = event.request.registry
-        updater = LocalUpdater(signer=registry.signers[key],
-                               storage=registry.storage,
-                               permission=registry.permission,
-                               source=resource['source'],
-                               destination=resource['destination'])
-
-        new_status = new_collection.get("status")
-        if new_status == "to-sign":
-            # Run signature process (will set `last_reviewer` field).
-            try:
-                updater.sign_and_update_destination(event.request)
-            except Exception:
-                logger.exception("Could not sign '{0}'".format(key))
-                event.request.response.status = 503
-
-        elif new_status == "to-review":
-            # Track `last_editor`
-            if "last_editor" not in new_collection:  # XXX why recursivity ?
-                updater.update_source_editor(event.request)
-
-
-def check_collection_status(event, resources, force_groups, force_review,
-                            editors_group, reviewers_group):
-    """Make sure status changes are allowed.
-    """
-    payload = event.payload
-
-    if 'bucket_id' not in payload:  # Pragma: no cover
-        # Safety check for kinto < 3.3 where events have incoherent payloads
-        # on default bucket.
-        return
-
-    editors_group = "/buckets/{bucket_id}/groups/{group_name}".format(
-        group_name=editors_group, **payload)
-    reviewers_group = "/buckets/{bucket_id}/groups/{group_name}".format(
-        group_name=reviewers_group, **payload)
-
-    current_user_id = event.request.prefixed_userid
-    current_principals = event.request.effective_principals
-
-    for impacted in event.impacted_records:
-        old_collection = impacted.get("old", {}).copy()
-        old_status = old_collection.get("status")
-        new_collection = impacted["new"].copy()
-        new_status = new_collection.get("status")
-
-        key = "/buckets/{bucket_id}/collections/{collection_id}".format(
-            bucket_id=payload["bucket_id"],
-            collection_id=new_collection["id"])
-
-        if key not in resources:
-            continue
-
-        if old_status == new_status:
-            continue
-
-        # Only these status can be set manually.
-        if new_status in ("signed",):
-            raise_invalid(message="Cannot set status to '%s'" % new_status)
-
-        # 1. None -> work-in-progress
-        # 2. work-in-progress -> to-review
-        elif new_status == "to-review":
-            if editors_group not in current_principals and force_groups:
-                raise_forbidden(message="Not in editors group")
-
-        # 3. to-review -> work-in-progress
-        elif new_status == "work-in-progress":
-            pass
-
-        # 3. to-review -> to-sign
-        elif new_status == "to-sign":
-            # Only allow to-sign from to-review if reviewer and no-editor
-            if reviewers_group not in current_principals and force_groups:
-                raise_forbidden(message="Not in reviewers group")
-
-            if old_status not in ("to-review", "signed") and force_review:
-                raise_invalid(message="Collection not reviewed")
-
-            if old_collection.get("last_editor") == current_user_id:
-                raise_forbidden(message="Editor cannot review")
-
-        # 4. to-sign -> signed
-
-        # Nobody can remove the status
-        elif new_status is None:
-            raise_invalid(message="Cannot remove status")
-        # Unknown manual status
-        else:
-            raise_invalid(message="Invalid status '%s'" % new_status)
-
-
-def check_collection_tracking(event, resources):
-    """Make sure tracking fields are not changed manually/removed.
-
-    XXX: Use readonly field notion from kinto.core ?
-    """
-    payload = event.payload
-
-    if 'bucket_id' not in payload:  # Pragma: no cover
-        # Safety check for kinto < 3.3 where events have incoherent payloads
-        # on default bucket.
-        return
-
-    tracking_fields = ("last_author", "last_editor", "last_reviewer")
-
-    for impacted in event.impacted_records:
-        old_collection = impacted.get("old", {})
-        new_collection = impacted["new"]
-
-        for field in tracking_fields:
-            old = old_collection.get(field)
-            new = new_collection.get(field)
-            if old != new:
-                raise_invalid(message="Cannot change %r" % field)
-
-
-def set_work_in_progress_status(event, resources):
-    """Put the status in work-in-progress if was signed.
-    """
-    payload = event.payload
-
-    if 'bucket_id' not in payload:  # Pragma: no cover
-        # Safety check for kinto < 3.3 where events have incoherent payloads
-        # on default bucket.
-        return
-
-    key = "/buckets/{bucket_id}/collections/{collection_id}".format(**payload)
-    resource = resources.get(key)
-    if resource is None:
-        return
-
-    registry = event.request.registry
-    updater = LocalUpdater(
-        signer=registry.signers[key],
-        storage=registry.storage,
-        permission=registry.permission,
-        source=resource['source'],
-        destination=resource['destination'])
-    updater.update_source_status("work-in-progress", event.request)
 
 
 def _signer_dotted_location(settings, resource):
@@ -267,13 +81,13 @@ def includeme(config):
                               reviewers_group=reviewers_group)
 
     config.add_subscriber(
-        functools.partial(set_work_in_progress_status, resources=resources),
+        functools.partial(listeners.set_work_in_progress_status,
+                          resources=resources),
         ResourceChanged,
-        for_resources=('record',)
-    )
+        for_resources=('record',))
 
     config.add_subscriber(
-        functools.partial(check_collection_status,
+        functools.partial(listeners.check_collection_status,
                           resources=resources,
                           force_review=force_review,
                           force_groups=force_groups,
@@ -281,19 +95,18 @@ def includeme(config):
                           reviewers_group=reviewers_group),
         ResourceChanged,
         for_actions=(ACTIONS.CREATE, ACTIONS.UPDATE),
-        for_resources=('collection',)
-    )
+        for_resources=('collection',))
 
     config.add_subscriber(
-        functools.partial(check_collection_tracking, resources=resources),
+        functools.partial(listeners.check_collection_tracking,
+                          resources=resources),
         ResourceChanged,
         for_actions=(ACTIONS.CREATE, ACTIONS.UPDATE),
-        for_resources=('collection',)
-    )
+        for_resources=('collection',))
 
     config.add_subscriber(
-        functools.partial(sign_collection_data, resources=resources),
+        functools.partial(listeners.sign_collection_data,
+                          resources=resources),
         ResourceChanged,
         for_actions=(ACTIONS.CREATE, ACTIONS.UPDATE),
-        for_resources=('collection',)
-    )
+        for_resources=('collection',))
